+++ R/clumping.R
@@ -70,11 +70,13 @@ snp_clumping <- function(G, infos.chr,
                          ncores = 1) {
 
   check_args()
+  assert_lengths(infos.chr, cols_along(G))
 
   if (!missing(is.size.in.bp))
     warning2("Parameter 'is.size.in.bp' is deprecated.")
 
-  if (!is.null(S)) assert_lengths(infos.chr, S)
+  if (!is.null(infos.pos)) assert_lengths(infos.pos, infos.chr)
+  if (!is.null(S)) assert_lengths(S, infos.chr)
 
   ind.noexcl <- setdiff(seq_along(infos.chr), exclude)
 
@@ -156,8 +158,8 @@ snp_pruning <- function(G, infos.chr,
 
 #' Long-range LD regions
 #'
-#' 34 long-range Linkage Disequilibrium (LD) regions for the human genome
-#' based on some [wiki table](https://goo.gl/0Ou7uI).
+#' 34 long-range Linkage Disequilibrium (LD) regions for the human genome based on
+#' [this wiki table](https://genome.sph.umich.edu/wiki/Regions_of_high_linkage_disequilibrium_(LD)).
 #'
 #' @format A data frame with 34 rows (regions) and 4 variables:
 #' - `Chr`: region's chromosome
@@ -167,8 +169,7 @@ snp_pruning <- function(G, infos.chr,
 "LD.wiki34"
 
 #' @param LD.regions A `data.frame` with columns "Chr", "Start" and "Stop".
-#' Default use the table of 34 long-range LD regions that you can find
-#' [there](https://goo.gl/0Ou7uI).
+#' Default use [LD.wiki34].
 #'
 #' @import foreach
 #' @export