+++ tests/testthat/test-2-autoSVD.R
@@ -10,6 +10,8 @@ CHR <- test$map$chromosome
 POS <- test$map$physical.pos / 10
 POS2 <- round(POS + 1)
 
+obj.bed <- bed(snp_writeBed(test, bedfile = tempfile(fileext = ".bed")))
+
 ################################################################################
 
 test_that("snp_autoSVD() works", {
@@ -17,10 +19,11 @@ test_that("snp_autoSVD() works", {
   expect_error(snp_autoSVD(G, as.character(CHR), POS), "only integers")
   expect_error(snp_autoSVD(G, CHR[-1], POS2), bigstatsr:::GET_ERROR_DIM())
   expect_error(snp_autoSVD(G, CHR, POS2[-1]), bigstatsr:::GET_ERROR_DIM())
+  expect_error(snp_autoSVD(G, CHR, min.mac = 0), "no variation; set min.mac > 0")
   expect_output(snp_autoSVD(G, CHR, POS2, thr.r2 = NA), "Skipping clumping.")
 
   expect_is(obj.svd <- snp_autoSVD(G, CHR, verbose = FALSE), "big_SVD")
-  expect_identical(attr(obj.svd, "lrldr"), LD.wiki34[0, 1:3])
+  expect_identical(attr(obj.svd, "lrldr"), cbind(LD.wiki34[0, 1:3], Iter = integer()))
 
   skip_if(is_cran)
 
@@ -41,6 +44,20 @@ test_that("snp_autoSVD() works", {
   obj.svd7 <- snp_autoSVD(G, CHR, alpha.tukey = 0.999, roll.size = 0, verbose = FALSE)
   expect_lt(length(attr(obj.svd7, "subset")), length(attr(obj.svd6, "subset")))
 
+  obj.svd8 <- snp_autoSVD(G, CHR, POS, alpha.tukey = 0.9999, roll.size = 0,
+                          int.min.size = 0, verbose = FALSE)
+  lrldr8 <- attr(obj.svd8, "lrldr")
+  expect_identical(sapply(lrldr8, typeof),
+                   c(Chr = "integer", Start = "double", Stop = "double", Iter = "integer"))
+  expect_gte(min(lrldr8$Iter), 1)
+
+  obj.svd9 <- snp_autoSVD(G, CHR + 0, test$map$physical.pos, size = 5000, alpha.tukey = 0.9999,
+                          roll.size = 0, int.min.size = 0, verbose = FALSE)
+  lrldr9 <- attr(obj.svd9, "lrldr")
+  expect_identical(sapply(lrldr9, typeof),
+                   c(Chr = "double", Start = "integer", Stop = "integer", Iter = "integer"))
+  expect_gte(min(lrldr9$Iter), 1)
+
   expect_output(
     snp_autoSVD(G, CHR, alpha.tukey = 0.999999999, roll.size = 0, verbose = TRUE),
     "Maximum number of iterations reached.")
@@ -50,12 +67,11 @@ test_that("snp_autoSVD() works", {
 
 test_that("bed_autoSVD() works", {
 
-  obj.bed <- bed(snp_writeBed(test, bedfile = tempfile(fileext = ".bed")))
-
+  expect_error(bed_autoSVD(obj.bed, min.mac = 0), "no variation; set min.mac > 0")
   expect_output(bed_autoSVD(obj.bed, thr.r2 = NA), "Skipping clumping.")
 
   expect_is(obj.svd <- bed_autoSVD(obj.bed, verbose = FALSE), "big_SVD")
-  expect_identical(attr(obj.svd, "lrldr"), LD.wiki34[0, 1:3])
+  expect_identical(attr(obj.svd, "lrldr"), cbind(LD.wiki34[0, 1:3], Iter = integer()))
 
   skip_if(is_cran)
 
@@ -77,6 +93,31 @@ test_that("bed_autoSVD() works", {
   expect_output(
     bed_autoSVD(obj.bed, alpha.tukey = 0.999999999, roll.size = 0, verbose = TRUE),
     "Maximum number of iterations reached.")
+})
+
+################################################################################
+
+test_that("MAC/MAF thresholds work", {
+
+  info <- bed_MAF(obj.bed)
+
+  replicate(10, {
+
+    min.mac <- sample(1:40, size = 1)
+    min.maf <- runif(n = 1, min = 0.01, max = 0.1)
+
+    obj.svd1 <- snp_autoSVD(G, CHR, size = 5, min.mac = min.mac, min.maf = min.maf,
+                            thr.r2 = NA, max.iter = 0, verbose = FALSE)
+    ind1 <- attr(obj.svd1, "subset")
+    expect_true(all(info$maf[ind1] >= min.maf))
+    expect_true(all(info$mac[ind1] >= min.mac))
+
+    obj.svd2 <- bed_autoSVD(obj.bed, min.mac = min.mac, min.maf = min.maf,
+                            thr.r2 = NA, max.iter = 0, verbose = FALSE)
+    ind2 <- attr(obj.svd2, "subset")
+    expect_true(all(info$maf[ind2] >= min.maf))
+    expect_true(all(info$mac[ind2] >= min.mac))
+  })
 })
 
 ################################################################################