+++ tests/testthat/test-4-pcadapt.R
@@ -30,7 +30,7 @@ test_that("Same as pcadapt", {
   test <- bigsnpr:::multLinReg(G, rows_along(G), cols_along(G), obj.svd$u, 1)
 
   expect_equal(obj.svd$center / 2, obj.pcadapt$af)
-  # TODO: update when new version of {pcadapt} is on CRAN
+  # TODO: change this test when new version of {pcadapt} on CRAN
   # expect_equal(obj.svd$d, obj.pcadapt$singular.values * sqrt((nrow(G) - 1) * ncol(G)))
   expect_equal(abs(obj.svd$u), abs(obj.pcadapt$scores))
   expect_equal(abs(obj.svd$v), abs(obj.pcadapt$loadings))